+++ src/main/java/functional/figure/figures/Pawn.java
@@ -1,6 +1,6 @@
 package functional.figure.figures;
 
-import functional.Board;
+import util.Board;
 import functional.figure.Figure;
 import functional.figure.PlayerDependentFigure;
 import functional.figure.special.KillLinked;
